Joueur PDO add/read

+++ Model/Club.php
@@ -1,12 +1,11 @@
 <?php 
-include "./bd/config.php";
-include"./Views/console.php";
+include_once "./bd/config.php";
+include_once "./Views/console.php";
 class Club {
     private $con, $id, $nameBd, $name, $ville, $created_at;
     function __construct(PDO $con)
     {
         $this->con = $con;
-
     }
 
     // name/getter/setter
@@ -56,7 +55,9 @@ class Club {
     public function getAll ()
     {
         $sql = "SELECT * FROM club ORDER BY id DESC";
-        return $this->con->query($sql)->fetchAll(PDO::FETCH_ASSOC);
+        
+        $leTouts=$this->con->query($sql)->fetchAll(PDO::FETCH_ASSOC);
+        return $leTouts;
     }
     //=============CREAT==========
     public function creatClub ()
@@ -68,24 +69,37 @@ class Club {
             ":name" => $this->name,
             ":ville" => $this->ville,
             ":created_at" => $this->created_at
-
         ]);
+    }
+    public function checkId(){
+        $sql ="SELECT id FROM club WHERE id = :id";
+
+        $stmt = $this->con->prepare($sql);
+        $stmt->execute([
+            ":id" => $this->id
+        ]);
+       return $stmt->fetch(PDO::FETCH_ASSOC);
     }
 
     //=============DELETE==========
     public function delete ()
     {
+    
+        if($this->checkId()){
         $sql = "DELETE FROM club WHERE id = :id";
         $stmt = $this->con->prepare($sql);
 
         return $stmt->execute([
             ":id" => $this->id
         ]);
+        }
+        echo "id n'exist pas";
     }
 
     //=============EDIT==========
     public function edit ()
     {
+    if($this->checkId()){
         $sql = "UPDATE club 
                 SET Name = :name, Ville = :ville, 
                 Date_Creation = :created_at
@@ -97,30 +111,9 @@ class Club {
     ":created_at" => $this->created_at,
     ":id" => $this->id
     ]);
-
-      
+    }
+    echo "id n'exist pas";
     }
 };
 
-// $club = new Club($con);
-// print_r($club->);
 
-
-$club = new Club($con);
-// $console = new Console();
-// echo "Nom du club: ";
-//     $name = $console->input("Entre votre choix");;
-
-// echo "Ville du club: ";
-//     $ville = $console->input("Entre votre choix");;
-
-// echo "Date de création (YYYY-MM-DD): ";
-//     $date = $console->input("Entre votre choix");
-
-// $club->setName($name);
-// $club->setVille($ville);
-// $club->setDate($date);
-// $club->setId(1);
-
- // $club->edit ();
-// $club->delete ();
